# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NURTW Membership and Vehicle Verification System",
    "version": "0.1.0",
    "description": "Internal administration API for the National Union of Road Transport Workers, Anambra State Council.\n\nA verification enquiry answered by this system confirms only that a matching NURTW record exists under the requested criteria. It is not evidence of ownership, roadworthiness, licensing, or insurance, and no response should be presented as such.\n\nThis document is generated from the running application: the route list is read from the dependency-injection container and request bodies from the same schemas the API validates against. It cannot describe a route that does not exist."
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nurtw_session",
        "description": "Opaque session token issued by POST /auth/login and set as an httpOnly cookie. Revocation takes effect on the next request."
      }
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/v1/organisations": {
      "get": {
        "summary": "Retrieve the organisational hierarchy.",
        "description": "Returns a forest, not a single tree. The response is filtered to the subtrees in which the caller holds `organisation.read`, so a branch administrator legitimately receives a view with no council at its head. The internal materialised path is never returned.\n\n**Permission required:** `organisation.read`.",
        "operationId": "OrganisationController_tree",
        "tags": [
          "Organisation"
        ],
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "401": {
            "description": "No valid session cookie was presented.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The session does not hold organisation.read in the relevant scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an organisation.",
        "description": "Authorised against the **parent’s** path: a node that does not yet exist has no path of its own, so the question is whether the caller may add beneath that parent. A council is created with `parentId: null` and therefore requires the permission at the root. Placement is exact — a zone sits beneath a council, a branch beneath a zone, a unit beneath a branch, and nothing skips a level.\n\n**Permission required:** `organisation.manage`.",
        "operationId": "OrganisationController_create",
        "tags": [
          "Organisation"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  "level": {
                    "type": "string",
                    "enum": [
                      "COUNCIL",
                      "ZONE",
                      "BRANCH",
                      "UNIT"
                    ]
                  },
                  "parentId": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "stateName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  }
                },
                "required": [
                  "name",
                  "level",
                  "parentId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created."
          },
          "400": {
            "description": "The body failed validation. Field-level detail is returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No valid session cookie was presented.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The session does not hold organisation.manage in the relevant scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "The parent does not exist, or lies outside the caller’s scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "The requested placement is not valid for that level, or the parent is inactive.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}